u-boot-rockpi: add dependency on radxa-binary-loader
On some build machines, the compilation of u-boot fails with error tools/mkimage
SPL size too large. The reason for that is that the mkimage cannot find the
ddr.bin at the expected DEPLOYDIR location due to race conditions. Therefore,
explicitly add radxa-binar-loader to DEPENDS to avoid race conditions and make
sure that ddr.bin is available at the expected location.

Signed-off-by: Yasir Khan <yasir_electronics@yahoo.com>